Remaining Visual C++ 2005 compile fixes

+++ lib/backupclient/BackupStoreDirectory.h
@@ -146,7 +146,8 @@ public:
 	int64_t GetRevisionID() const {return mRevisionID;}
 	void SetRevisionID(int64_t RevisionID) {mRevisionID = RevisionID;}
 	
-	unsigned int GetNumberOfEntries() const {return mEntries.size();}
+	unsigned int GetNumberOfEntries() const 
+		{return (unsigned int)mEntries.size();}
 
 	// User info -- not serialised into streams
 	int64_t GetUserInfo1_SizeInBlocks() const {return mUserInfo1;}
